+++ src/Mod/Points/App/Properties.h
@@ -155,7 +155,7 @@ private:
 /** Curvature information. */
 struct PointsExport CurvatureInfo
 {
-    double fMaxCurvature, fMinCurvature;
+    float fMaxCurvature, fMinCurvature;
     Base::Vector3f cMaxCurvDir, cMinCurvDir;
 };
 
@@ -182,7 +182,7 @@ public:
     int getSize(void) const {return _lValueList.size();}   
     void setValue(const CurvatureInfo&);
     void setValues(const std::vector<CurvatureInfo>&);
-    std::vector<double> getCurvature( int tMode) const;
+    std::vector<float> getCurvature( int tMode) const;
 
     /// index operator
     const CurvatureInfo& operator[] (const int idx) const {return _lValueList.operator[] (idx);}
